+++ src/test/java/wooteco/subway/dao/StationDaoTest.java
@@ -5,21 +5,18 @@ import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
-import javax.sql.DataSource;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
-import org.springframework.jdbc.core.JdbcTemplate;
 
 import wooteco.subway.domain.Station;
 
 @JdbcTest
-@Import(JdbcStationDao.class)
+@Import(StationDao.class)
 class StationDaoTest {
     private static final String STATION_NAME = "청구역";
